backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
chromadb